% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobody.R
\name{aden}
\alias{aden}
\title{Brings forth the product}
\usage{
aden(x)
}
\arguments{
\item{x}{an array of numbers}
}
\value{
a product of all the members
}
\description{
Brings forth the product
}
\examples{
prod(runif(10))
}
