% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clusterLouvain}
\alias{clusterLouvain}
\title{Cluster Cells via Louvain Algorithm}
\usage{
clusterLouvain(
  nn_network,
  python_path = NULL,
  resolution = 1,
  weight_col = NULL,
  louv_random = F,
  set_seed = T,
  seed_number = 0,
  ...
)
}
\arguments{
\item{nn_network}{Constructed nearest neighbor network to use}

\item{python_path}{Specify specific path to python if required}

\item{resolution}{Resolution}

\item{weight_col}{Weight column}

\item{louv_random}{Random}

\item{set_seed}{Set seed}

\item{seed_number}{Number for seed}

\item{...}{Additional parameters}
}
\value{
A character vector of cluster labels
}
\description{
Cluster cells using a NN-network and the Louvain algorithm from the community module in Python
}
