% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{constructNN}
\alias{constructNN}
\title{Nearest Network}
\usage{
constructNN(
  reduced_object,
  k_neighbors = 30,
  minimum_shared = 5,
  top_shared = 3,
  verbose = F,
  ...
)
}
\arguments{
\item{reduced_object}{PC reduction matrix}

\item{k_neighbors}{Number of k neighbors to use}

\item{minimum_shared}{Minimum shared neighbors}

\item{top_shared}{Keep at ...}

\item{verbose}{Be verbose}

\item{\dots}{Additional parameters}
}
\value{
NN network as igraph object
}
\description{
Construct a nearest neighbour network based on previously computed PCs
}
