% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bolivia.R
\name{jack}
\alias{jack}
\title{Finds the sum of two values}
\usage{
jack(a, b)
}
\arguments{
\item{a}{first value}

\item{b}{second value}
}
\value{
a sum of the first and second values
}
\description{
Finds the sum of two values
}
\examples{
jack(2, 3)
}
