% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDIGO_scanner.R
\encoding{UTF-8}
\name{KDIGO_SCr_scanner}
\alias{KDIGO_SCr_scanner}
\title{Implementation of the KDIGO AKI guideline SCr criteria for AKI diagnosis and staging}
\usage{
KDIGO_SCr_scanner(data, wrong_stage_3_crit = FALSE)
}
\arguments{
\item{data}{Data frame with one SCr sample per row and the following columns: Admission IDs ('admission' - Integer), admission time ('admission_time' - POSIX), SCr sample timestamps ('time' - POSIX) and SCr sample values ('value' - Numeric) in μmol/L}

\item{wrong_stage_3_crit}{Boolean, if TRUE, an incorrect stage 3 SCr criterion will be used to stage admissions with AKI: An increase in SCr by 4.0 mg/dl (353.6 μmol/L) instead of an increase to or beyond 4.0 mg/dl (353.6 μmol/L)}
}
\value{
Data frame containing the original data and the added AKI diagnosis and stage status per SCr sample:

[,1] - 'admission': admission identification number (integer)

[,2] - 'admission_time': admission timestamp (POSIX)

[,3] - 'time': timestamp of SCr value (POSIX)

[,4] - 'value': SCr concentration (μmol/L, numeric)

[,5] - 'hours_since_admission': number of hours passed since the admission (numeric)

[,6] - 'creatinine_baseline': SCr baseline concentration, defined as the first SCr within the first 24h of admission (μmol/L ,numeric)

[,7] - 'creatinine_increase_large': any increase >= 26.5 μmol/L during the admission (logical)

[,8] - 'max_creatinine_increase_previous_48': max increase in previous 48h. Only calculated for non-baseline values if creatinine_increase_large == TRUE (μmol/L, numeric)

[,9] - 'AKI_absolute_increase': set to 1 if max_creatinine_increase_previous_48 >= 26.5 μmol/L, else set to 0. Set to NA for baseline values (numeric; [0, 1, NA]). See the KDIGO AKI guideline at 2.1.1 (page 19)

[,10] - 'AKI_relative_increase': set to 1 if SCr value / baseline >= 1.5 and increase occurred within previous 7 days, else set to 0. Set to NA for baseline values (integer; [0, 1, NA]). See the KDIGO AKI guideline at 2.1.1 (page 19)

[,11] - 'AKI': set to 1 if AKI_absolute_increase or AKI_relative_increase is set to 1. Set to NA for baseline values (integer; [0, 1, NA])

[,12] - 'stage_abs_incr': absolute increase in SCr for AKI cases. Calculated for SCr samples at or after AKI onset, by subtracting the minimum SCr before AKI onset from the SCr sample value (μmol/L, numeric)

[,13] - 'stage_rel_incr': relative increase in SCr for AKI cases. Calculated for SCr samples at or after AKI onset, by dividing the SCr sample value by the SCr baseline (numeric)

[,14] - 'AKI_stage_abs': AKI stage associated with the absolute increase in SCr and the absolute SCr value according to the KDIGO AKI SCr criteria (numeric; [1, 2, 3, NA]). See the KDIGO AKI guideline at 2.1.2, Table 2 (page 19) and the "Timeframe for diagnosis and staging" paragraph on pages 30-31

[,15] - 'AKI_stage_rel': AKI stage associated with the relative increase in SCr according to the KDIGO AKI SCr criteria (numeric; [1, 2, 3, NA]). See the KDIGO AKI guideline at 2.1.2, Table 2 (page 19) and the "Timeframe for diagnosis and staging" paragraph on pages 30-31

[,16] - 'AKI_stage': Maximum stage from the absolute increase stage and the relative increase stage (numeric; [1, 2, 3, NA])
}
\description{
Acute kidney injury (AKI) diagnosis and staging using the Kidney Disease: Improving Global Outcomes (KDIGO) AKI serum creatinine (SCr) guideline criteria [1]. Admissions are first screened for AKI; AKI is diagnosed if the SCr sample values show an absolute or relative (compared to the SCr baseline) increase within the required time-windows, as specified in the KDIGO AKI guideline (See the KDIGO AKI guideline at 2.1.1 on page 19). Next, admissions with AKI are staged; for each SCr sample value at and after AKI onset, the absolute increase (compared to the minimum value before AKI onset), relative increase (compared to the SCr baseline) and absolute SCr values are computed and associated stages assigned. See the KDIGO AKI guideline at 2.1.2, Table 2 (page 19) and the "Timeframe for diagnosis and staging" paragraph on pages 30-31.

All data presented in the KDIGO SCr scanner dummy dataset are entirely fictional and perhaps even nonsensical. They are meant to serve an educational purpose in understanding how the code for AKI diagnosis and staging according to KDIGO AKI guideline criteria, works. Furthermore, while very effort was made to ensure that accurate information was implemented in the code to implement AKI diagnosis and staging according to the KIDIGO AKI guideline serum creatinine criteria, we accept no liability or responsibility to any person or organization as a consequence of any reliance upon the information contained in the code or code use.

[1] Kellum JA, Lameire N, Aspelin P, et al. Kidney disease: Improving global outcomes (KDIGO) acute kidney injury work group. KDIGO clinical practice guideline for acute kidney injury. Kidney International Supplements. 2012;2(1):1-138.

https://kdigo.org/wp-content/uploads/2016/10/KDIGO-2012-AKI-Guideline-English.pdf
}
