% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDIGO_SCr_scanner_dummy_data.R
\docType{data}
\name{KDIGO_SCr_scanner_dummy_data}
\alias{KDIGO_SCr_scanner_dummy_data}
\title{Dummy dataset to test the KDIGO_SCr_scanner function}
\format{
A data frame with 297 observations and 4 variables:

[,1] - 'admission': admission identification number (integer)

[,2] - 'admission_time': admission timestamp (POSIX)

[,3] - 'time': timestamp of serum creatinine value (POSIX)

[,4] - 'value': serum creatinine concentration (μmol/L, numeric)
}
\usage{
KDIGO_SCr_scanner_dummy_data
}
\description{
Dummy dataset to test the KDIGO_SCr_scanner function
}
\keyword{datasets}
