% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bolivia.R
\name{last}
\alias{last}
\title{finds the mean of an array}
\usage{
last(x)
}
\arguments{
\item{x}{a numerical input}
}
\value{
the final value
}
\description{
finds the mean of an array
}
\examples{
last(runif(10))
}
