% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{mod.get_de_genes}
\alias{mod.get_de_genes}
\title{Modified get_de_genes() function from RCTD (Cable et al., Nature Biotechnology, 2022)}
\usage{
mod.get_de_genes(
  cell_type_info,
  puck,
  fc_thresh = 0.75,
  expr_thresh = 2e-04,
  MIN_OBS = 10
)
}
\arguments{
\item{cell_type_info}{cell type information and profiles of each cell, calculated from the scRNA-seq reference (see \code{\link{get_cell_type_info}})}

\item{fc_thresh}{minimum \code{log_e} fold change required for a gene (Default: 0.75)}

\item{expr_thresh}{minimum expression threshold, as normalized expression (Default: 2e-04)}

\item{MIN_OBS}{the minimum number of occurrences of each gene in Y.count (Default: 10)}

\item{Y.count}{Gene x Grid (or Spot) raw count matrix}
}
\value{
\item{total_gene_list}{a list of differntially expressed gene names}
}
\description{
Modified get_de_genes() function from RCTD (Cable et al., Nature Biotechnology, 2022)
}
