% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{mod.remove_platform_effect}
\alias{mod.remove_platform_effect}
\title{Modified remove_platform_effect() function from RCTD (Cable et al., Nature Biotechnology, 2022)}
\usage{
mod.remove_platform_effect(Y.count, X.count, pseudocount = 1)
}
\arguments{
\item{Y.count}{Gene x Grid (or Spot) raw count matrix}

\item{X.count}{Gene x Cell raw count matrix from reference data}

\item{pseudocount}{pseudo-counts to be added (Default: 1)}
}
\value{
\item{X.count_corrected}{Gene x Cell platform-effect-normalized count matrix}
}
\description{
Modified remove_platform_effect() function from RCTD (Cable et al., Nature Biotechnology, 2022)
}
