% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{nnls.naive}
\alias{nnls.naive}
\title{Naive Non-negative Least Squares (NNLS)}
\usage{
nnls.naive(Y, X, ncores = NULL, chunk.size = NULL)
}
\arguments{
\item{Y}{Gene x Grid (or Spot) cpm-normalized count matrix}

\item{X}{Gene x Feature matrix: either (1) average expression matrix or (2) NMF feature matrix}

\item{ncores}{number of cores to use}

\item{chunk.size}{number of chunk to split}
}
\value{
\item{alpha.hat}{estimated cell type proportion matrix (Feature x Grid)}
\item{resid}{estimated residual value matrix (Gene x Grid)}
\item{resid.truncated}{non-negative truncated residual value matrix (Gene x Grid)}
}
\description{
Y_\{G x J\} ~= X_\{G x K\} alpha_\{K x J\} + R_\{G x J\}
}
