% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESCUE.R
\name{RESCUE}
\alias{RESCUE}
\title{RESCUE}
\usage{
RESCUE(
  Y.count,
  X.count,
  X.cluster,
  C.grid = NULL,
  seed = 1,
  max.iter = 100,
  eps = 1e-04,
  platform_norm = TRUE,
  fc_thresh = 0.75,
  expr_thresh = 2e-04,
  MIN_OBS = 10,
  CELL_MIN_INSTANCE = 25,
  ncores = NULL,
  chunk.size = NULL,
  verbose = FALSE
)
}
\arguments{
\item{Y.count}{Gene x Grid (or Spot) raw count matrix}

\item{X.count}{Gene x Cell raw count matrix from reference data}

\item{X.cluster}{vector of cell type annotations for each cell from reference data}

\item{C.grid}{vector of tuning parameters; Default: seq(from=1e-6, to=5, length=10)}

\item{seed}{random seed for generating Poisson surrogate model (Default: 1)}

\item{max.iter}{maximum number of iterations (Default: 100)}

\item{eps}{convergence tolerance (Default: 1e-4)}

\item{platform_norm}{if true, apply platform effect normalization (Default: TRUE)}

\item{fc_thresh}{minimum log-fold-change (across cell types) in gene filtering (Default: 0.75)}

\item{expr_thresh}{minimum normalized gene expression in gene filtering (Default: 2e-4)}

\item{MIN_OBS}{minimum total counts per pixel of genes (Default: 10)}

\item{ncores}{number of cores to use; Default: detectCores()-1}

\item{chunk.size}{number of chunk to split; Default: round(ncol(Y.count) / 100)}

\item{CELL_MIN_INSANCE}{minimum number of cells requires per cell type (Default: 25)}
}
\value{
\item{alpha.hat}{estimated cell type proportion matrix (Feature x Grid)}
\item{R.hat}{estimated residual matrix in CPM scale (Gene x Grid)}
\item{L.hat}{estimated cellness matrix in CPM scale (Gene x Grid)}
\item{R.raw}{estimated residual matrix in raw scale (Gene x Grid)}
\item{L.raw}{estimated cellness matrix in raw scale (Gene x Grid)}
\item{rel}{vector of relative errors}
\item{C.grid}{vector of tuning parameters}
}
\description{
Y_\{G x J\} ~= X_\{G x K\} alpha_\{K x J\} + R_\{G x J\}
}
