% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{wnn.ser}
\alias{wnn.ser}
\title{Weighted Non-negative Sparse Error Recovery}
\usage{
wnn.ser(
  Y,
  X,
  Y.nUMI,
  R.init = NULL,
  C = 2,
  max.iter = 100,
  eps = 1e-04,
  verbose = FALSE
)
}
\arguments{
\item{Y}{Gene x Grid (or Spot) cpm-normalized count matrix}

\item{X}{Gene x Feature matrix: either (1) average expression matrix or (2) NMF feature matrix}

\item{Y.nUMI}{vector of UMI counts of unnormalized Y}

\item{C}{tuning parameter}

\item{max.iter}{maximum number of iterations}

\item{eps}{convergence tolerance}
}
\value{
\item{loss}{vector of loss function values at each iteration}
\item{w}{estimated weight matrix (Gene x Grid)}
\item{alpha.hat}{estimated cell type proportion matrix (Feature x Grid)}
\item{R.hat}{estimated residual matrix (Gene x Grid)}
\item{L.hat}{estimated cellness matrix (Gene x Grid)}
}
\description{
Y_\{G x J\} ~= X_\{G x K\} alpha_\{K x J\} + R_\{G x J\}
}
