% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{sampleImputation}
\alias{sampleImputation}
\title{Sample-mean Estimation}
\usage{
sampleImputation(
  expression_matrix,
  subset_genes = NULL,
  scale_data = TRUE,
  number_pcs = 8,
  snn_resolution = 0.9,
  impute_index = NULL,
  pseudo_zero = NULL,
  python_path = NULL,
  verbose = FALSE
)
}
\arguments{
\item{expression_matrix}{Row by column log-normalized expression matrix}

\item{subset_genes}{A vector of informative gene names, defaults to all genes}

\item{scale_data}{Whether to standardize expression by gene, default TRUE}

\item{number_pcs}{Number of dimensions to inform SNN clustering}

\item{snn_resolution}{Resolution parameter for SNN}

\item{impute_index}{Index to impute, will default to all zeroes}

\item{pseudo_zero}{Pseudo-zero expression value}

\item{python_path}{path to your python binary (default = system path)}

\item{verbose}{Print progress output to the console}
}
\value{
Returns a sparse matrix of class 'dgCMatrix'
}
\description{
Cluster cells using SNN and a list of given genes, estimate missing expression
values for each cell-gene combination with the within-cluster non-zero expression
mean
}
\examples{
\dontrun{
sampleImputation(expression_matrix)
}

}
